chore(suite): router thunks
