+++ suite/router/package.json
@@ -13,6 +13,8 @@
     },
     "dependencies": {
         "@reduxjs/toolkit": "2.11.2",
+        "@suite/locks": "workspace:*",
+        "@suite/modal": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
         "@suite-common/suite-config": "workspace:*",
         "@suite-common/suite-types": "workspace:*",
