+++ suite/router/src/index.ts
@@ -5,5 +5,6 @@ export * from './createSuiteRouterHistory';
 export * from './routerMiddleware';
 export * from './routerParams';
 export * from './routerReducer';
+export * from './routerThunks';
 export * from './routes';
 export * from './suiteRouterHistory';
